VMWare: Disabling linked clone doesn't cache images
This fixes an issue when use_linked_clone is false and the image is
already in the cache.  The problem was that the image would be
copied from glance into the instance directory even if it was already
in the cache.  This change checks if the file is in the cache, only
copies from glance if it is not, and copies the image from the cache
into the instance directory if use_linked_clone is false.

Fixes: bug #1207064
(cherry picked from commit c53b0f66802ed82fea8c637f0138c0a01e159efa)

Conflicts:

	nova/tests/virt/vmwareapi/test_vmwareapi.py
	nova/virt/vmwareapi/vmops.py

Change-Id: I635e00448fc014b3841ed43b590d451305e8ff2f
(cherry picked from commit 2f94155c34209422ce9c0c3d643e48aa6b367054)