VMware ESX: Boot from volume must not relocate vol
ESX does not support relocate API. Moreover when working with ESX the boot
from volume logic need not perform relocation since the volume's datastore
is always visible to the ESX host.

The fix is to ignore relocation when working with ESX.

Also made a trivial logging change to read task name only when provided
by the server. Task name is optional property and will not always be set by
the server. This is to avoid AttributeError which was constantly observed
when instance was being powered-on.

Change-Id: I227cf2ea5b6f681150ede0bf6257eb238f1452d2
Closes-bug: #1244918

Conflicts:

	nova/tests/virt/vmwareapi/test_vmwareapi.py