VMware: upload images to temporary directory
The patch ensures that new images are uploaded to a defined
temporary directory instead of directly to the cache directory.

The reason for this are as follows:
1. More than one nova-compute may try and upload the same image
to the datastore at the same time
2. If there are errors then the temporary folder can be purged
(this will be done in a different patch)
3. When the image is downloaded it will be moved to the cache
directory. Only once an image has been successfully downloaded
will it be moved. This will help with false positives with
cache image deletion

Two new configuration variables have been added (in the 'vmware'
folder):
    - tmp_folder - the temporary folder
    - cache_folder - the folder to cache images

The first phase of blueprint vmware-image-cache-management.

DocImpact

Change-Id: I4f3ee90ae866583e8df782ff89d2b11bbebf9c39

Conflicts:

	nova/virt/vmwareapi/vmops.py